Start of backend redesign
This is the start if the redesign of the MMH backend for MobHeads. Eventually all heads will be in json files, either flat or nested.
Mobs like Creaking would be flat, while Villagers would be nested. Nested mobs can have a defaultChance so that Admins would not need to set all entries if they are the same.